Streamline assembly of the system matrix.
The assembled system matrix and RHS vector are used for direct linear solvers in the context of implicit integration.
When invoked (currently only from the MKL solver), the ConvertToMatrixForm function must always fill in both the matrix and the RHS vector and include all constraints in the system.
As such, all unnecessary checks were removed.