Change return type for functions returning violations for body-mesh constraints.
Returning a ChMatrix<> by value is not correct.  Objects of this type should not be created directly (and ChMatrix<> does not have a proper copy constructor).  The implementation of ChMatrix<> needs to be revisited to properly impose this.  Right now, since *no* constructor is defined, a default constructor is automatically generated!